requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.4
billiard==4.2.0
boto3==1.35.4
botocore==1.35.4
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
dill==0.3.8
Django==5.0.8
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
injector==0.22.0
isort==5.13.2
jmespath==1.0.1
kombu==5.4.0
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycoingecko==3.1.0
PyJWT==2.9.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.3.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
